lcs-sacoche:trunk update to version 2011-06-25
git-svn-id: https://svn.tice.ac-caen.fr/svn/LcsSE3/LCS/LCS2.0ndg/sources/trunk@4022 1012e355-7652-49d3-8fd6-1f9b0b464cea

+++ lcs-sacoche/Sacoche/_sql/structure/sacoche_socle_entree.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS sacoche_socle_entree;
 
 CREATE TABLE sacoche_socle_entree (
 	entree_id    SMALLINT(5)  UNSIGNED                NOT NULL AUTO_INCREMENT,
-	section_id   SMALLINT(5)  UNSIGNED                NOT NULL DEFAULT 0,
+	section_id   TINYINT(3)   UNSIGNED                NOT NULL DEFAULT 0,
 	entree_ordre TINYINT(3)   UNSIGNED                NOT NULL DEFAULT 0 COMMENT "Commence à 0.",
 	entree_nom   VARCHAR(235) COLLATE utf8_unicode_ci NOT NULL DEFAULT "",
 	PRIMARY KEY (entree_id),
